Fix: Review changes

+++ src/core/settings.h
@@ -90,7 +90,7 @@ typedef struct QUIC_SETTINGS_INTERNAL {
     uint16_t MaxBindingStatelessOperations;
     uint16_t StatelessOperationExpirationMs;
     uint16_t CongestionControlAlgorithm;
-    uint64_t IdleSrcCidChangeMs;
+    uint32_t IdleSrcCidChangeMs;
 
 } QUIC_SETTINGS_INTERNAL;
